refactor: simplify issue5.js and update test structure
