+++ test/fixtures/issue5.js
@@ -1,14 +1,5 @@
 'use strict'
 
-const int = setInterval(() => {}, 10)
-process.on('message', () => {
-  process._rawDebug('!!! doit')
-  clearInterval(int)
-  replication()
-})
-
-function replication () {
-  const pino = require('pino')()
-  pino.info('hello world')
-  setTimeout(() => { throw new Error('uncaught') }, 20)
-}
+const pino = require('pino')()
+pino.info('hello world')
+setTimeout(() => { throw new Error('uncaught') }, 20)
