+++ test/utils.js
@@ -46,8 +46,19 @@ function createSecureTcpListener (msgHandler) {
   })
 }
 
+function withResolvers () {
+  let _resolve
+  let _reject
+  const promise = new Promise((resolve, reject) => {
+    _resolve = resolve
+    _reject = reject
+  })
+  return { promise, resolve: _resolve, reject: _reject }
+}
+
 module.exports = {
   createSecureTcpListener,
   createTcpListener,
-  createUdpListener
+  createUdpListener,
+  withResolvers
 }
